Use gulp-inline-source for everything
Previously, these steps were happening:

- Inline the "critical" CSS
- Concatenate CSS and JS into single files respectively
- Minify those files
- Rename those files to keep their path unique
- Update the link in the HTML

However, the "critical" CSS was many thousands of lines.

As this is a single-page (well, two-page - England and China) application,
I figured that basically all of the CSS and JavaScript is pretty
critical, so why not inline it all?

I found the module "gulp-inline-source", which can also conpress the CSS
and JS at the same time, so I figured, why not just do it all at once?
This means we don't need to worry about keeping the links unique, or
concatenating into 1 file or any of it.

So that's what I've done.

Not to mention that this almost halves the size of node_modules,
reducing by about 70MB.